Update table.tcl
work in progress. format is going to just be a matrix, where you index the rows (starting at 1, since 0 is the header), and the column indices are stored in a hashmap with fieldnames for quick lookup. The only requirements for this new implementation are that the first row has unique names. I am moving away from the double-nested dictionary, it took up way too much space, was far too complex, and led to shimmering problems.